pyproject.toml
[project]
authors = [{name = "Yogesh Thambidurai", email = "yogesh@gyoge.com"}]
name = "sub_volume"
requires-python = ">= 3.11"
version = "0.1.0"
dependencies = [
  "funlib-geometry>=0.3.0,<0.4",
  "glfw>=2.9.0,<3",
  "numpy",
  "pygfx",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
sub_volume = { path = ".", editable = true }

[tool.pixi.tasks]
main = "python scripts/main.py"

[tool.pixi.dependencies]
numpy = ">=2.3.0,<3"
zarr = ">=3.0.8,<4"
pygfx = ">=0.12.0,<0.13"
tensorstore = ">=0.1.65,<0.2"

[tool.pixi.feature.test.dependencies]
pytest = ">=8.4.1,<9.0"
hypothesis = ">=6.135.16,<7"

[tool.pixi.feature.dev.dependencies]
pixi-pycharm = ">=0.0.8,<0.0.9"
ruff = ">=0.11.10, <0.12"

[tool.pixi.feature.test.tasks]
test = "pytest ."

[tool.pixi.feature.dev.tasks]
# F401: unused imports, I: isort
# both of these should count as formatting for us
format = "ruff check --select F401,I --fix . && ruff format ."
check = "ruff check ."

[tool.pixi.environments]
dev = ["dev", "test"]

[tool.ruff.lint]
preview = true
pydocstyle.convention = "google"
select = [
    # ruff defaults
    "E4", "E7", "E9", "F",
    # imports
    "F401", "I",
    # simplify statements
    "SIM",
    # pytest
    "PT",
    # comprehensions
    "C4",
    # builtins
    "A",
    # unused arguments
    "ARG",
    # numpy
    "NPY",
    # naming conventions
    "N",
    # extraneous docstring return/yield/raise
    # also consider adding DOC as a whole?
    "DOC202", "DOC403", "DOC502",
    # general docstring stuff
    "D", 
    # no blank lines before class docstrings
    "D211",
    # multiline docstrings start on second line
    "D213",
    # python upgrades
    "UP",
]

ignore = [
    # conflicts with D211 and D213
    "D203", "D212",
]

[tool.ruff.lint.per-file-ignores]
"**/{tests,scripts}/*" = ["D", "DOC"]
